[SW-258] Revert extractCSS config
