+++ examples/basic-examples/nuxt.config.js
@@ -34,7 +34,6 @@ export default {
     '@/assets/_custom.scss'
   ],
   build: {
-    extractCSS: true,
     extend (config, { isDev, isClient }) {
       if (isDev) {
         // For debugging in dev mode
