+++ packages/create-ripple-app/generators/updates/1.23.1/saofile.js
@@ -1,6 +1,6 @@
 const common = require('./../../common')
 const templateDir = './../../../template'
-const jscodeshift = require('jscodeshift')
+// const jscodeshift = require('jscodeshift')
 
 module.exports = {
   ...common,
@@ -16,34 +16,35 @@ module.exports = {
           data.devDependencies['@dpc-sdp/ripple-test-tools'] = `1.9.0`
           return data
         }
-      },
-      {
-        type: 'modify',
-        files: 'nuxt.config.js',
-        handler (data, filepath) {
-          /*
-          // https://astexplorer.net/
-          // example of using jscodeshift for changing the site value and not modifying anything else
-          const j = jscodeshift
-          const defaultNode = j(data).find(j.ExportDefaultDeclaration)
-          const nuxtConfig = defaultNode.nodes()[0].declaration.properties
-          const tide = nuxtConfig.filter(e => e.key.name === 'tide')
-          const site = tide[0].value.properties.filter(p => p.key.name === 'site')[0]
-          site.value.value = 123
-          const out = defaultNode.toSource({ quote: 'single' })
-          return out
-          */
-          const j = jscodeshift
-          const rootData = j(data)
-
-          const defaultNode = rootData.find(j.ExportDefaultDeclaration)
-          defaultNode.find(j.Property, { key: { name: 'build' } })
-            .find(j.Property, { key: { name: 'extend' } })
-            .insertAfter(`extractCSS: { ignoreOrder: true }`)
-          const out = rootData.toSource({ quote: 'single' })
-          return out
-        }
       }
+      // Below script is disabled by SW-258
+      // {
+      //   type: 'modify',
+      //   files: 'nuxt.config.js',
+      //   handler (data, filepath) {
+      //     /*
+      //     // https://astexplorer.net/
+      //     // example of using jscodeshift for changing the site value and not modifying anything else
+      //     const j = jscodeshift
+      //     const defaultNode = j(data).find(j.ExportDefaultDeclaration)
+      //     const nuxtConfig = defaultNode.nodes()[0].declaration.properties
+      //     const tide = nuxtConfig.filter(e => e.key.name === 'tide')
+      //     const site = tide[0].value.properties.filter(p => p.key.name === 'site')[0]
+      //     site.value.value = 123
+      //     const out = defaultNode.toSource({ quote: 'single' })
+      //     return out
+      //     */
+      //     const j = jscodeshift
+      //     const rootData = j(data)
+
+      //     const defaultNode = rootData.find(j.ExportDefaultDeclaration)
+      //     defaultNode.find(j.Property, { key: { name: 'build' } })
+      //       .find(j.Property, { key: { name: 'extend' } })
+      //       .insertAfter(`extractCSS: { ignoreOrder: true }`)
+      //     const out = rootData.toSource({ quote: 'single' })
+      //     return out
+      //   }
+      // }
     ]
     return actions
   },
