+++ packages/create-ripple-app/generators/updates/1.26.0/saofile.js
@@ -1,5 +1,6 @@
 const common = require('./../../common')
 const templateDir = './../../../template'
+const jscodeshift = require('jscodeshift')
 
 module.exports = {
   ...common,
@@ -13,6 +14,19 @@ module.exports = {
           data = data.replace(' style="mix-blend-mode: overlay;"', '')
           return data
         }
+      },
+      {
+        type: 'modify',
+        files: 'nuxt.config.js',
+        handler (data, filepath) {
+          const j = jscodeshift
+          const rootData = j(data)
+
+          const defaultNode = rootData.find(j.ExportDefaultDeclaration)
+          defaultNode.find(j.Property, { key: { name: 'extractCSS' } }).remove()
+          const out = rootData.toSource({ quote: 'single' })
+          return out
+        }
       }
     ]
     return actions
@@ -20,5 +34,6 @@ module.exports = {
   async completed () {
     console.log('SDP 1.26.0:')
     console.log('[SDPA-4339] Fixed oops banner')
+    console.log('[SW-258] Fixed extractCSS issue')
   }
 }
